AI_ASSISTANT/requirements.txt
PyQt6
SpeechRecognition
pyaudio
edge-tts
google-generativeai
ollama
opencv-python-headless
pygame
